requirements.txt
streamlit

openai

langchain-0.0.271

langchain-experimental

pyodbc

pandas